tests for task statuses

+++ app/Http/Controllers/TaskStatusController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreTaskStatusRequest;
 use App\Http\Requests\UpdateTaskStatusRequest;
 use App\Models\TaskStatus;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 
@@ -22,18 +23,24 @@ class TaskStatusController extends Controller
 
     /**
      * @return View
+     * @throws AuthorizationException
      */
     public function create(): View
     {
+        $this->authorize('create', TaskStatus::class);
+
         return view('taskStatus.create');
     }
 
     /**
      * @param StoreTaskStatusRequest $request
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(StoreTaskStatusRequest $request): RedirectResponse
     {
+        $this->authorize('create', TaskStatus::class);
+
         $taskStatus = new TaskStatus();
         $taskStatus->name = $request->name;
         $taskStatus->save();
@@ -46,9 +53,12 @@ class TaskStatusController extends Controller
     /**
      * @param TaskStatus $taskStatus
      * @return View
+     * @throws AuthorizationException
      */
     public function edit(TaskStatus $taskStatus): View
     {
+        $this->authorize('update', TaskStatus::class);
+
         return view('taskStatus.edit', [
             'taskStatus' => $taskStatus,
         ]);
@@ -58,9 +68,12 @@ class TaskStatusController extends Controller
      * @param UpdateTaskStatusRequest $request
      * @param TaskStatus $taskStatus
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(UpdateTaskStatusRequest $request, TaskStatus $taskStatus): RedirectResponse
     {
+        $this->authorize('update', TaskStatus::class);
+
         $taskStatus->name = $request->name;
         $taskStatus->save();
 
@@ -72,9 +85,12 @@ class TaskStatusController extends Controller
     /**
      * @param TaskStatus $taskStatus
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function destroy(TaskStatus $taskStatus): RedirectResponse
     {
+        $this->authorize('delete', TaskStatus::class);
+
         $taskStatus->delete();
 
         flash(__('flash.task-status.destroy'))->success();
